refactor:调整会话初始化逻辑

+++ mistchat-server/src/main/java/cc/xmist/mistchat/server/chat/resp/ContactResponse.java
@@ -21,6 +21,6 @@ public class ContactResponse {
         private Long chatId;
         private Long readMsgId;
         private Long lastMsgId;
-        private Integer unReadCount;
+        private Long unReadCount;
     }
 }
